+++ src/utils/SyntaxRuleError.ts
@@ -1,0 +1,13 @@
+const syntaxErrorMessages = [
+  "STATEMENT_WITHOUT_BOOLEAN",
+  "STATEMENT_WITH_INVALID_OPERATOR",
+  "STATEMENT_WITH_STANDALONE_NOT_OPERATOR",
+] as const;
+export type SyntaxRuleErrorMessage = (typeof syntaxErrorMessages)[number];
+
+export class SyntaxRuleError extends Error {
+  constructor(message: SyntaxRuleErrorMessage) {
+    super(message);
+    this.name = "SyntaxRuleError";
+  }
+}
